# apps/api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/demo/notes": {
      "get": {
        "operationId": "NoteController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "查询成功"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "便签列表",
        "tags": [
          "演示便签"
        ]
      },
      "post": {
        "operationId": "NoteController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "创建便签",
        "tags": [
          "演示便签"
        ]
      }
    },
    "/api/demo/notes/{id}": {
      "get": {
        "operationId": "NoteController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "便签详情",
        "tags": [
          "演示便签"
        ]
      },
      "patch": {
        "operationId": "NoteController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "更新便签",
        "tags": [
          "演示便签"
        ]
      },
      "delete": {
        "operationId": "NoteController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "删除便签",
        "tags": [
          "演示便签"
        ]
      }
    },
...
  },
  "info": {
    "title": "Admin Zen",
    "description": "Admin Zen API 文档",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "在 Authorization 头中传入 access token，格式：Bearer <token>"
      }
    },
...
  }
}